moved providers to own folder, removed project-provider
