+++ app/layout.tsx
@@ -1,22 +1,21 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { Header } from '@/components/header';
-import { Footer } from '@/components/footer';
-import { ProjectsProvider } from '@/components/projects-provider';
-import { ThemeProvider } from '@/components/theme-provider';
-import { cn } from '@/lib/utils'
-import { siteMetadata } from '@/data/metadata'
-import { professionalServiceSchema } from '@/data/structured-data'
+import { Header } from "@/components/Header";
+import { Footer } from "@/components/Footer";
+import { ThemeProvider } from "@/providers/theme-provider";
+import { cn } from "@/lib/utils";
+import { siteMetadata } from "@/data/metadata";
+import { professionalServiceSchema } from "@/data/structured-data";
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ["latin"] });
 
-export const metadata: Metadata = siteMetadata
+export const metadata: Metadata = siteMetadata;
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en" suppressHydrationWarning>
@@ -24,28 +23,24 @@ export default function RootLayout({
         <script
           type="application/ld+json"
           dangerouslySetInnerHTML={{
-            __html: JSON.stringify(professionalServiceSchema)
+            __html: JSON.stringify(professionalServiceSchema),
           }}
         />
       </head>
-      <body className={cn(inter.className, 'min-h-screen bg-background')}>
+      <body className={cn(inter.className, "min-h-screen bg-background")}>
         <ThemeProvider
           attribute="class"
           defaultTheme="system"
           enableSystem
           disableTransitionOnChange
         >
-          <ProjectsProvider>
-            <div className="flex min-h-screen flex-col">
-              <Header />
-              <main className="flex-1">
-                {children}
-              </main>
-              <Footer />
-            </div>
-          </ProjectsProvider>
+          <div className="flex min-h-screen flex-col">
+            <Header />
+            <main className="flex-1">{children}</main>
+            <Footer />
+          </div>
         </ThemeProvider>
       </body>
     </html>
-  )
+  );
 }
